add warning for overwrite

+++ evaluation/benchmarks/swe_bench/scripts/eval/combine_final_completions.py
@@ -32,7 +32,7 @@ def load_completions(output_dir: str, instance_id: str):
 
 
 parser = argparse.ArgumentParser()
-parser.add_argument('jsonl-path', type=str, required=True)
+parser.add_argument('jsonl_path', type=str)
 args = parser.parse_args()
 
 output_dir = os.path.dirname(args.jsonl_path)
@@ -40,10 +40,14 @@ df = pd.read_json(args.jsonl_path, lines=True, orient='records')
 df['raw_completions'] = df['instance_id'].progress_apply(
     lambda x: load_completions(output_dir, x)
 )
-
 print(f'Successfully loaded {len(df)} completions')
 
 output_path = os.path.join(output_dir, 'output.with_completions.jsonl')
+if os.path.exists(output_path):
+    print(f'Output file already exists at {output_path}, overwriting? (y/n)')
+    if input() != 'y':
+        print('Exiting...')
+        exit(0)
 # save to jsonl
 df.to_json(output_path, lines=True, orient='records')
 print(f'Saved to {output_path}')
